requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Collection
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF Generation
weasyprint==60.1
jinja2==3.1.2

# Data Visualization
matplotlib==3.8.2
numpy==1.26.2

# Database
aiosqlite==0.19.0

# Security & OAuth
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0

# Email & HTTP
httpx==0.25.2
aiofiles==23.2.1

# Environment
python-dotenv==1.0.0

# AI (for paid tier)
anthropic==0.7.8
